requirements.txt
bottle
# ical also looks good but more stars wins (sorry!)
icalendar
requests
waitress
pillow
